Some major updates, mainly to the editor
Fonts now render and are selectable. Images are changable as well
however the UI for it is a little ugly.

Most aspects of Text is complete, minus Rotation values

Images are still highly broken, you can't edit the width/height, and
selecting a new image will not fix these values.

Shapes need a few more options